V116

+++ tests/test_tl.py
@@ -29,8 +29,10 @@ def hto_and_gex():
     )
 
     obs = pd.DataFrame(index=cell_ids)
-    adata_hto = ad.AnnData(X=x_hto, dtype=x_hto.dtype, obs=obs)
-    adata_gex = ad.AnnData(X=x_gex, dtype=x_gex.dtype, obs=obs)
+    adata_hto = ad.AnnData(X=x_hto, obs=obs)
+    adata_gex = ad.AnnData(X=x_gex, obs=obs)
+    adata_hto.X = adata_hto.X.astype(x_hto.dtype)
+    adata_gex.X = adata_gex.X.astype(x_gex.dtype)
 
     return adata_hto, adata_gex
 
@@ -43,10 +45,10 @@ def hto_and_raw():
     x_hto_raw = np.ones((3, 3))
 
     obs = pd.DataFrame(index=cell_ids)
-    adata_hto = ad.AnnData(X=x_hto, dtype=x_hto.dtype, obs=obs.iloc[:3])
-    adata_hto_raw = ad.AnnData(
-        X=np.concatenate([x_hto, x_hto_raw]), dtype=x_hto.dtype, obs=obs
-    )
+    adata_hto = ad.AnnData(X=x_hto, obs=obs.iloc[:3])
+    adata_hto_raw = ad.AnnData(X=np.concatenate([x_hto, x_hto_raw]), obs=obs)
+    adata_hto.X = adata_hto.X.astype(x_hto.dtype)
+    adata_hto_raw.X = adata_hto_raw.X.astype(x_hto.dtype)
 
     return adata_hto, adata_hto_raw
